requirements.txt
PyPDF2
langchain
pinecone-client
openai
tiktoken
boto3